requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
streamlit
requests
llama-cpp-python
